+++ src/entities/piece/factory/class.ts
@@ -1,9 +1,9 @@
 import * as THREE from 'three'
 import { Piece } from '../single'
-import { PieceType } from '../single/types'
 import { initialCells } from '@shared/configs/pieces/initialCells'
-import { createPieceName } from '../single/utils'
 import { PlayerColor } from '@shared/configs/player/color'
+import { PieceType } from '@shared/configs/pieces/types'
+import { createPieceName } from '@shared/utils/pieces/createPieceName'
 
 export class PieceFactory {
     static create(type: PieceType, color: PlayerColor, index: number = 1, object: THREE.Group | THREE.Mesh): Piece {
