+++ src/entities/piece/single/index.ts
@@ -1,2 +1,1 @@
-export * from './class'
-export * from './types'
+export * from './base'
