+++ src/shared/configs/pieces/types/configs.ts
@@ -1,0 +1,8 @@
+export enum PieceType {
+    KING = 'king',
+    QUEEN = 'queen',
+    ROOK = 'rook',
+    BISHOP = 'bishop',
+    KNIGHT = 'knight',
+    PAWN = 'pawn'
+}
