+++ src/shared/types/piece/index.ts
@@ -1,2 +1,1 @@
 export * from './types'
-export * from './class'
