+++ src/shared/utils/pieces/createPieceName/utils.ts
@@ -1,5 +1,5 @@
+import { PieceType } from '@shared/configs/pieces/types'
 import { PlayerColor } from '@shared/configs/player/color'
-import { PieceType } from './types'
 
 export function createPieceName(color: PlayerColor, type: PieceType, index: number) {
     return `${color.charAt(0).toUpperCase() + color.slice(1)}${type.charAt(0).toUpperCase() + type.slice(1)}${index}`
